electrical1/requirement.txt
arabic-reshaper==2.1.3
asgiref==3.5.0
beautifulsoup4==4.10.0
bleach==4.1.0
bootstrap-admin==0.4.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
coupon-codes==0.1
cryptography==36.0.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.2
django-allauth==0.48.0
django-appconf==1.0.5
django-baton==2.2.3
django-blog-zinnia==0.20
django-ckeditor==6.2.0
django-compressor==3.1
django-contrib-comments==2.2.0
django-environ==0.8.1
django-facebook==6.0.3
django-filter==21.1
django-import-export==2.7.1
django-jazzmin==2.4.9
django-jet-reboot==1.2.0
django-js-asset==2.0.0
django-libsass==0.9
django-mptt==0.13.4
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-social-auth==0.7.28
django-summernote==0.8.20.0
django-tagging==0.5.0
django-xmlrpc==0.1.8
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.0.0
et-xmlfile==1.1.0
future==0.18.2
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
googleapis-common-protos==1.54.0
html5lib==1.1
httplib2==0.20.4
idna==3.3
libsass==0.21.0
MarkupPy==1.14
mots-vides==2015.5.11
mysql-connector-python==8.0.28
mysqlclient==2.1.0
oauth2==1.9.0.post1
oauthlib==3.2.0
odfpy==1.4.1
openpyxl==3.0.9
packaging==21.3
pdfkit==1.0.0
Pillow==9.0.1
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==1.7.1
pyparsing==3.0.7
PyPDF2==1.26.0
python-bidi==0.4.2
python-openid==2.2.5
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2021.3
PyYAML==6.0
rcssmin==1.1.0
regex==2022.1.18
reportlab==3.6.6
requests==2.27.1
requests-oauthlib==1.3.1
rjsmin==1.2.0
rsa==4.8
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
soupsieve==2.3.1
sqlparse==0.4.2
tablib==3.2.0
tzdata==2021.5
Unidecode==1.3.2
uritemplate==4.1.1
urllib3==1.26.8
webencodings==0.5.1
xhtml2pdf==0.2.5
xlrd==2.0.1
xlwt==1.3.0
